# Conflicts:
#	src/main/java/com/solegendary/reignofnether/ability/Ability.java
#	src/main/java/com/solegendary/reignofnether/ability/EnchantAbilityServerboundPacket.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/Bloodlust.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/ConsumeSlime.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/SpinWebs.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/ThrowHealingPotion.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/ThrowLingeringHarmingPotion.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/ThrowLingeringRegenPotion.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/ThrowWaterPotion.java
#	src/main/java/com/solegendary/reignofnether/building/Building.java
#	src/main/java/com/solegendary/reignofnether/building/BuildingSaveData.java
#	src/main/java/com/solegendary/reignofnether/building/ProductionBuilding.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/monsters/Dungeon.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/monsters/Graveyard.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/monsters/Laboratory.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/monsters/Mausoleum.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/monsters/SculkCatalyst.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/monsters/SlimePit.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/monsters/SpiderLair.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/monsters/Stronghold.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/neutral/Beacon.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/neutral/EndPortal.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/BasaltSprings.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/Bastion.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/CentralPortal.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/FlameSanctuary.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/Fortress.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/HoglinStables.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/Portal.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/WitherShrine.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/shared/AbstractStockpile.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/villagers/ArcaneTower.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/villagers/Barracks.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/villagers/Blacksmith.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/villagers/Castle.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/villagers/Library.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/villagers/TownCentre.java
#	src/main/java/com/solegendary/reignofnether/hud/HudClientEvents.java
#	src/main/java/com/solegendary/reignofnether/hud/PortraitRendererUnit.java
#	src/main/java/com/solegendary/reignofnether/hud/buttons/ActionButtons.java
#	src/main/java/com/solegendary/reignofnether/hud/buttons/HelperButtons.java
#	src/main/java/com/solegendary/reignofnether/player/PlayerClientEvents.java
#	src/main/java/com/solegendary/reignofnether/player/PlayerServerEvents.java
#	src/main/java/com/solegendary/reignofnether/registrars/PacketHandler.java
#	src/main/java/com/solegendary/reignofnether/registrars/ServerEventRegistrar.java
#	src/main/java/com/solegendary/reignofnether/research/ResearchServerboundPacket.java
#	src/main/java/com/solegendary/reignofnether/unit/UnitActionItem.java
#	src/main/java/com/solegendary/reignofnether/unit/units/monsters/CreeperUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/monsters/NecromancerUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/monsters/SkeletonUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/monsters/SlimeUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/monsters/SpiderUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/monsters/StrayUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/monsters/WardenUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/monsters/ZoglinUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/monsters/ZombieVillagerUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/neutral/EndermanUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/piglins/BlazeUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/piglins/BruteUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/piglins/GhastUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/piglins/GhastUnitFireball.java
#	src/main/java/com/solegendary/reignofnether/unit/units/piglins/GruntUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/piglins/HeadhunterUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/piglins/HoglinUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/piglins/PiglinMerchantUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/piglins/WitherSkeletonUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/villagers/EvokerUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/villagers/MilitiaUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/villagers/PillagerUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/villagers/RavagerUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/villagers/RoyalGuardUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/villagers/VillagerUnit.java
#	src/main/java/com/solegendary/reignofnether/unit/units/villagers/WitchUnit.java
#	src/main/java/com/solegendary/reignofnether/util/MiscUtil.java
#	src/main/java/com/solegendary/reignofnether/votesystem/MapSelectionScreen.java
#	src/main/java/com/solegendary/reignofnether/votesystem/VoteCommand.java

+++ src/main/java/com/solegendary/reignofnether/unit/goals/GenericTargetedSpellGoal.java
@@ -23,15 +23,15 @@ public class GenericTargetedSpellGoal extends MoveToTargetBlockGoal {
     protected final int channelTicks; // max time required to cast a spell
     protected boolean isCasting = false;
     protected BlockPos castTarget = null; // pos that the spell will be cast at
-    protected final int range;
+    protected final float range;
     public Consumer<LivingEntity> onEntityCast;
     public Consumer<BlockPos> onGroundCast;
     public Consumer<BuildingPlacement> onBuildingCast;
     public boolean hasKeyframeAnimations = false;
-    protected int bonusChannelingRange = 0; // extra range added while channeling
+    protected float bonusChannelingRange = 0; // extra range added while channeling
     protected final UnitAnimationAction animationStart;
 
-    public GenericTargetedSpellGoal(Mob mob, int channelTicks, int range,
+    public GenericTargetedSpellGoal(Mob mob, int channelTicks, float range,
                                     UnitAnimationAction animationStart,
                                     @Nullable Consumer<LivingEntity> onEntityCast,
                                     @Nullable Consumer<BlockPos> onGroundCast,
@@ -45,7 +45,7 @@ public class GenericTargetedSpellGoal extends MoveToTargetBlockGoal {
         this.onBuildingCast = onBuildingCast;
     }
 
-    public GenericTargetedSpellGoal(Mob mob, int channelTicks, int range,
+    public GenericTargetedSpellGoal(Mob mob, int channelTicks, float range,
                                     @Nullable Consumer<LivingEntity> onEntityCast,
                                     @Nullable Consumer<BlockPos> onGroundCast,
                                     @Nullable Consumer<BuildingPlacement> onBuildingCast) {
@@ -82,7 +82,7 @@ public class GenericTargetedSpellGoal extends MoveToTargetBlockGoal {
     }
 
     protected boolean isInRange() {
-        int finalRange = range;
+        float finalRange = range;
         if (isCasting())
             finalRange += bonusChannelingRange;
